+++ src/main/java/login/tikichat/domain/attachment/model/Attachment.java
@@ -10,54 +10,27 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Inheritance(strategy = InheritanceType.JOINED)
+@DiscriminatorColumn(name = "type")
 @Table(name = "attachments")
 public class Attachment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    protected Long id;
 
     @Column(length = 4000, nullable = false)
-    private String path;
+    protected String path;
 
     @Column(length = 1000, nullable = false)
-    private String filename;
+    protected String filename;
 
     @Column(length = 1000, nullable = false)
-    private String originalFilename;
+    protected String originalFilename;
 
     @Column(nullable = false)
-    private String ext;
+    protected String ext;
 
     @ManyToOne()
     @JoinColumn(name = "uploader_user_id", nullable = false)
-    private User uploaderUser;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "chat_room_id", nullable = false)
-    private ChatRoom chatRoom;
-
-    public Attachment(
-            User uploaderUser,
-            ChatRoom chatRoom,
-            String path,
-            String filename,
-            String originalFilename,
-            String ext
-    ) {
-        this.filename = filename;
-        this.uploaderUser = uploaderUser;
-        this.chatRoom = chatRoom;
-        this.path = path;
-        this.originalFilename = originalFilename;
-        this.ext = ext;
-    }
-
-    // ChatRoom 과의 연관관계 설정 편의 메서드
-    public void setChatRoom(ChatRoom chatRoom) {
-        this.chatRoom = chatRoom;
-
-        if (chatRoom != null && !chatRoom.getAttachments().contains(this)) {
-            chatRoom.getAttachments().add(this);
-        }
-    }
+    protected User uploaderUser;
 }
