+++ src/main/java/login/tikichat/domain/chatroom/service/ChatRoomService.java
@@ -1,6 +1,7 @@
 package login.tikichat.domain.chatroom.service;
 
 import login.tikichat.domain.attachment.model.Attachment;
+import login.tikichat.domain.attachment.model.ChatRoomAttachment;
 import login.tikichat.domain.category.dto.FindCategoryDto;
 import login.tikichat.domain.category.repository.CategoryRepository;
 import login.tikichat.domain.chatroom.constants.ChatRoomSortType;
@@ -147,7 +148,7 @@ public class ChatRoomService {
     }
 
     @Transactional
-    public void linkAttachment(Long chatRoomId, Attachment attachment) {
+    public void linkAttachment(Long chatRoomId, ChatRoomAttachment attachment) {
         ChatRoom chatRoom = chatRoomRepository.findById(chatRoomId).orElseThrow(ChatRoomNotFoundException::new);
         chatRoom.addAttachment(attachment);
     }
